requirements.txt
Django==2.1.15
djangorestframework==3.11.1
djongo==1.3.1
django-cors-headers==3.2.1
